Slight changes to the README, fixed buggy choice code.
I edited the README slightly to reflect that I am going to be staying
with the current Python 3 version (as of this commit, it is 3.3.3).
In the code, I made the getPlayerColor() function a little cleaner
and made a new getComputerColor() function for some separation of
tasks, and a more verbose codebase.

    - David Hagerty