Made a printable game board.
The game board is now fully printed at the start of the game. To do
this I made an array containing a representation of all the pieces on
the board. In addition, there are now variables for how many pieces
both players have remaining. These will be used for the checkWin()
function. Overall progress is going well.

    - David